# api-contract.yaml
openapi: "3.1.0"
info:
  title: PDF Processing SaaS API
  version: "1.0.0"
  description: API for uploading, processing, and downloading PDF files.

servers:
  - url: http://localhost:8000
    description: Local development

paths:
  /upload:
    post:
      operationId: uploadFile
      summary: Upload a file for processing
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUploadResponse"
        "413":
          description: File too large (max 50MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid file type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs:
    post:
      operationId: createJob
      summary: Create a PDF processing job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreateRequest"
      responses:
        "200":
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobCreateResponse"
        "404":
          description: One or more file_ids not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid request (bad tool, wrong file count)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{id}:
    get:
      operationId: getJobStatus
      summary: Get job status and details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /download/{id}:
    get:
      operationId: downloadResult
      summary: Download processed file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 'attachment; filename="output.pdf"'
        "400":
          description: Job not completed yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /files/{id}:
    delete:
      operationId: deleteFile
      summary: Delete an uploaded file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /health:
    get:
      operationId: healthCheck
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

components:
  schemas:
    FileUploadResponse:
      type: object
      required: [file_id, filename, size]
      properties:
        file_id:
          type: string
          description: UUIDv4 hex (32 chars, no dashes)
          example: "a1b2c3d4e5f6789012345678abcdef01"
        filename:
          type: string
          example: "document.pdf"
        size:
          type: integer
          description: File size in bytes
          example: 1048576

    JobCreateRequest:
      type: object
      required: [tool, file_ids]
      properties:
        tool:
          type: string
          enum: [merge, split, compress, convert, edit]
        file_ids:
          type: array
          items:
            type: string
          minItems: 1
        options:
          type: object
          default: {}
          description: "Tool-specific options. split: {page_ranges: '1-3,5'}. convert: {convert_to: 'docx'|'pdf'}. edit: {operations: [...]}. merge/compress: {}."

    JobCreateResponse:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
          description: UUIDv4 hex (32 chars, no dashes)
        status:
          type: string
          example: "queued"

    JobDetail:
      type: object
      required: [job_id, tool, status, progress, created_at, updated_at]
      properties:
        job_id:
          type: string
        tool:
          type: string
          enum: [merge, split, compress, convert, edit]
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        output_path:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "File not found"